LayoutTests/fast/dom/NamedNodeMap-named-getter-caching-expected.txt

Tests caching of result of NamedNodeMap named property getter

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bodyAttributes.class is undefined
PASS lastIterationHasRightValue is true
PASS successfullyParsed is true

TEST COMPLETE
